setting keyboard LEDs locks up the keyboard, so don't
doubtless I'm doing something wrong but I can't spot it.

Here's what I get on the keyboard. I get the same with the PIO code
running in CircuitPython, so if it's a bug it's in both places (no surprise
as the PIO code is essentially identical), but then how does initialization
work right?

> ED (pico -> keyboard)
< FA (keyboard -> pico)
> 02
< F8 (but maybe a bit has been miscounted and it's FE, resend?)