Introduce command keystrokes & add cyclable settings
now if you switch to UART1 and put a jumper or small-value
resistor between GP0 and GP1 you get echo back from typing.

UART2 (pins 12/13) not tested.

Confusingly UART1/2 both use the uart0 peripheral. ssssh.. this is
to keep uart0 as the connection to the DCE and uart1 as the connection
to the future daisy chained device such as a printer or punch.